server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
psycopg[binary]==3.1.19
psycopg-pool==3.2.2
pydantic==2.8.2
python-dotenv==1.0.1

